+++ msbuild/Xamarin.Shared/Xamarin.Shared.targets
@@ -1070,6 +1070,7 @@ Copyright (C) 2018 Microsoft. All rights reserved.
 			Condition="'$(IsMacEnabled)' == 'true'"
 			DeviceSpecificIntermediateOutputPath="$(DeviceSpecificIntermediateOutputPath)"
 			IsWatchApp="$(IsWatchApp)"
+			MaxDegreeOfParallelism="$(ColladaMaxDegreeOfParallelism)"
 			ProjectDir="$(MSBuildProjectDirectory)"
 			ResourcePrefix="$(_ResourcePrefix)"
 			SdkPlatform="$(_SdkPlatform)"
@@ -2771,6 +2772,7 @@ Copyright (C) 2018 Microsoft. All rights reserved.
 			Condition="'$(IsMacEnabled)' == 'true'"
 			Executable="$(_AppContainerDir)%(_NativeStripItems.Identity)"
 			Kind="%(_NativeStripItems.Kind)"
+			MaxDegreeOfParallelism="$(SymbolStripMaxDegreeOfParallelism)"
 			SymbolFile="%(_NativeStripItems.SymbolFile)"
 		/>
